♻️ [Refactoring]: #301 milestone ドメイン型を domains/milestone へ移し依存方向を lib/tauri → domains に統一
MilestoneDefinition / MilestoneState の定義を domains/milestone/index.ts へ移し、
domains 層が IPC 層（@/lib/tauri）を import していた双方向依存を解消する。
lib/tauri/milestoneCommands/types.ts は domains 型を import + 再公開し、
taskCommands/types.ts → @/domains/priority と同じ「lib/tauri → domains」方向に揃える。
IPC payload 形状と各関数シグネチャは不変（動作変更なし）。
